Refactor commands into seperate modules
