+++ apistar/commands/__init__.py
@@ -1,3 +1,5 @@
-from apistar.commands.base import new, run, test
+from apistar.commands.new import new
+from apistar.commands.run import run
+from apistar.commands.test import test
 
 __all__ = ['new', 'run', 'test']
